Move kubectl to the internal package (#137)
* PMM-5988 Fix order of size and ready fields (#134)

* Move kubectl to the internal package
